Put the line_sender in a separate task, since fuse() made it stop working
> With this change, almost all the original code forked from rustyline-async
> is now gone! That code was confusing with so many libraries thrown into the
> mix and undefined behaviors, from mixing and matching things from different
> crates! I've cleaned up Cargo.toml! Just tokio FTW!

In Readline::readline(), the main loop has to use `fuse()` due to
crossterm's EventStream. This means that only the 1st control signal
sent will be recv()'d. The rest are subsequently ignored.

This change moves line_sender to its own task. And moves it out of
the main event loop started by Readline::readline(). This can now
be used as a way to control the Readline and SharedWriter(s) whether
Readline::readline() is running or not.

Also remove all the deps on non-tokio async crates. And clean up
imports.

Remove:
- futures-channel = "0.3"
- futures-util = { version = "0.3", features = ["io"] }

Fixes: #28